修改数据库字段类型、

+++ imking-web/src/main/java/io/imking/reward/beans/dto/RwDetailDto.java
@@ -19,7 +19,7 @@ public class RwDetailDto {
 	
 	private int level;
 	
-	private BigDecimal taskAmount;
+	private Integer taskAmount;
 	
 	private String title;
