+++ imking-web/src/main/java/io/imking/reward/domain/RwProAct.java
@@ -1,7 +1,7 @@
 package io.imking.reward.domain;
 
-import java.math.BigDecimal;
 import java.util.Date;
+
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
@@ -20,7 +20,7 @@ public class RwProAct {
 
     private Byte type;
 
-    private BigDecimal amount;
+    private Integer amount;
 
     private Byte applyDelayDays;
